+++ src/prism_eda/analysis/__init__.py
@@ -2,6 +2,7 @@
 
 from prism_eda.analysis.anomaly import anomaly_detection_dataset
 from prism_eda.analysis.classification import classification_dataset
+from prism_eda.analysis.image_profile import profile_image_dataset
 from prism_eda.analysis.profile import profile_dataset
 from prism_eda.analysis.schema_discovery import discover_schema_dataset
 
@@ -9,5 +10,6 @@ __all__ = [
     "anomaly_detection_dataset",
     "classification_dataset",
     "discover_schema_dataset",
+    "profile_image_dataset",
     "profile_dataset",
 ]